[mod] mod uint32_t -> size_t

+++ inc/functions.h
@@ -6,7 +6,7 @@
 /*   By: shogura <shogura@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/30 19:11:59 by tharaguc          #+#    #+#             */
-/*   Updated: 2022/07/21 11:57:49 by shogura          ###   ########.fr       */
+/*   Updated: 2022/07/21 12:08:44 by shogura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,18 +33,15 @@ char	*extract_arg(t_shell *data, char *input, char **start, size_t *i);
 char	*store_quoted_arg(t_shell *data, char *input, size_t *i, char quote);
 char	*expand_env(char *arg, t_shell *data);
 char	*create_expanded_arg(char *arg, t_list **val, size_t len);
-char	**split_by_pipe(char *input, uint32_t cmd_cnt);
+char	**split_by_pipe(char *input, size_t cmd_cnt);
 size_t	count_arg_len(char *arg, t_list **val, t_list **key);
 size_t	count_args(char *input);
-u_int32_t	count_cmds(char *input);
+size_t	count_cmds(char *input);
 void	copy_env_val(char **dst, size_t *j, t_list **env_val);
 void	store_redirect_in_out(t_shell *data, char *input);
 
-
-//execution
-
-//executor
-void	executor(t_shell *shell);
+// executor
+void executor(t_shell *shell);
 int		ft_execvp(char *file, char *argv[]);
 
 #endif
